Updated research materials with a little more digging I did on item weights and costs. Added a few more pictures for main menu and final points screen for reference purposes. Worked on the interfaces and classes related to items and vehicles, realizing the store class and systems were pretty neglected but that is because I had not got this far yet but now that I am here jeeezzzz...

+++ Src/TrailCommon/Entity Templates/Vehicle/IVehicle.cs
@@ -10,7 +10,7 @@ namespace TrailCommon
     {
         IGameSimulation CurrentGame { get; }
         ReadOnlyCollection<IItem> Inventory { get; }
-        uint Balance { get; }
+        float Balance { get; }
         ReadOnlyCollection<IPerson> People { get; }
         RationLevel Ration { get; }
         TravelPace Pace { get; }
