+++ Src/TrailCommon/Modes/Store/IStoreMode.cs
@@ -6,7 +6,7 @@ namespace TrailCommon
     {
         ReadOnlyCollection<IItem> StoreInventory { get; }
         string StoreName { get; }
-        uint StoreBalance { get; }
+        float StoreBalance { get; }
         void BuyItems(IItem item);
         void SellItem(IItem item);
     }
